third_party/blink/web_tests/http/tests/xmlhttprequest/cross-site-denied-response-sync-2-expected.txt

CONSOLE ERROR: Access to XMLHttpRequest at 'http://localhost:8000/xmlhttprequest/resources/reply.xml' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
PASS
